feat(protocol,server): outbound MP_REACH + dual-stack prefix pipeline — phase 4b of #14
Sources go dual-stack: PrefixCidr gains a family-aware TryParse (IPv4 1..32,
IPv6 1..128, /0 rejected, v4-mapped rejected) and PrefixListParser emits
family-tagged prefixes — the old "SkipsIpv6" behavior flips per the epic.
RIPEstat parses v6.originating alongside v4.originating. The provider
pipeline (SourceLoadResult, source/user caches, RIPEstat cache) carries
IpPrefix; the Contracts boundary (IPrefixService) carries
(UInt128 Prefix, byte Length, bool IsIpv4[, uint Asn]) tuples — Contracts is
a dependency-free leaf and cannot see the Protocol type. RouteAssembler and
RouteSeedingService build family-tagged routes.

The send path splits by family: IPv4 rides classic NLRI + NEXT_HOP exactly
as before; IPv6 rides MP_REACH_NLRI (RFC 4760 §5 — no classic NEXT_HOP, next
hop inside the attribute, RFC 2545 §3 global form) and only when the peer
negotiated MP IPv6/Unicast AND Bgp.NextHopIpv6 is configured (validated as
global unicast 2000::/3). Otherwise IPv6 routes are suppressed with a
warning — fail-visible, never silently. Withdrawals mirror the split:
IPv4 via the withdrawn field, IPv6 via MP_UNREACH_NLRI (RFC 4760 §7).

7 wire-level session tests (scripted connection + real reader): MP_REACH
announce with no classic NEXT_HOP, reader round-trip, suppression without
negotiation/next-hop config, MP_UNREACH-only withdrawal on the wire,
attribute flag checks, config validation. 995/995 green; integration stand
exercises the real path against BIRD2 (2001:db8:cccc::/64 seed advertised
via MP_REACH over the IPv6-transport session).

New config: Bgp.NextHopIpv6 (optional; unset = no IPv6 advertisements —
IPv4-only deployments need no changes).

+++ BGPLite.Tests/BgpSessionRouteOwnershipTests.cs
@@ -189,9 +189,9 @@ public class BgpSessionRouteOwnershipTests
     }
 
     /// <summary>Waits for the initial dump and returns every NLRI the session put on the wire.</summary>
-    private static async Task<List<(uint Prefix, byte Length)>> CollectAdvertisedNlriAsync(ScriptedConnection conn)
+    private static async Task<List<(UInt128 Prefix, byte Length)>> CollectAdvertisedNlriAsync(ScriptedConnection conn)
     {
-        var nlri = new List<(uint, byte)>();
+        var nlri = new List<(UInt128, byte)>();
         for (var i = 0; i < 200; i++)
         {
             nlri.Clear();
@@ -386,7 +386,7 @@ public class BgpSessionRouteOwnershipTests
 
     // ---- harness ----
 
-    private static RouteTable Seeded(params (uint Prefix, byte Length)[] routes)
+    private static RouteTable Seeded(params (UInt128 Prefix, byte Length)[] routes)
     {
         var table = new RouteTable();
         foreach (var (prefix, length) in routes)
